Refactor financial calculations to use `BigDecimal` for precision and update database entity mappings.
- **Financial Precision**: Migrated multiple DTOs and Services (including `CartResponseDTO`, `MedicineDTO`, `PreOrderDTO`, and `BucketServiceImpl`) from `double` to `BigDecimal` to ensure accurate monetary calculations and rounding.
- **Entity & Repository Updates**:
    - Renamed `Stock` entity table mapping from `stocks` to `prices` and updated column mappings for `price_id` and `expiry_date`.
    - Updated `Vendor` entity to map to `vendor_informations`.
    - Adjusted `Medicine` entity with explicit column definitions and `LONGTEXT` constraints for descriptive fields.
- **Clean-up**: Removed obsolete classes including `DataSeeder`, `OtcProduct`, `OtcProductDTO`, `IOtcProductService`, `OtcProductServiceImpl`, `OtcProductRepository`, `Substitute`, and `SubstituteRepository`.
- **Logic Improvements**: Updated `MainServiceImpl` and `PreOrderServiceImpl` to handle `BigDecimal` arithmetic and improved stock/discount calculation logic.

+++ src/main/java/com/trego/dto/MedicineDTO.java
@@ -3,6 +3,7 @@ package com.trego.dto;
 import com.trego.dao.entity.Stock;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -39,13 +40,13 @@ public class MedicineDTO {
 
     private List<Stock> offLineStocks;
     private List<Stock> onLineStocks;
-    private double mrp;
-    private double discount;
+    private BigDecimal mrp;
+    private BigDecimal discount;
     private int qty;
     private String expiryDate;
-    private Double actualPrice;
+    private BigDecimal actualPrice;
     private String image;
     private String strip;
-    private Double offeredPrice;
+    private BigDecimal offeredPrice;
     private Long salesCount; // Added field for sales count
 }
